+++ electron/filesystem.ts
@@ -305,6 +305,7 @@ export class SettingsStore implements Settings.Store {
             const file = path.join(getSettingsDir(character), key);
             return <Settings.Keys[K]>JSON.parse(fs.readFileSync(file, 'utf8'));
         } catch(e) {
+            console.error('READ KEY FAILURE', e, key, character);
             return undefined;
         }
     }